refactor: isolate SQLite preview query generation
