+++ src/infra/adapters/sqlite/sql/metadata.rs
@@ -1,6 +1,6 @@
 use crate::app::ports::outbound::{DbOperationError, SQLITE_TABLE_LIST_REQUIRED_MARKER};
 
-use super::literal::{encode_preview_column_expr, quote_ident, quote_literal};
+use super::literal::{quote_ident, quote_literal};
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub(in crate::adapters::sqlite) enum TableMetadataQueryMode {
@@ -79,45 +79,6 @@ pub(in crate::adapters::sqlite) fn table_list_required_error() -> DbOperationErr
 
 pub(in crate::adapters::sqlite) fn is_table_list_unavailable(error: &str) -> bool {
     error.to_ascii_lowercase().contains("pragma_table_list")
-}
-
-pub(in crate::adapters::sqlite) fn build_preview_query(
-    table: &str,
-    columns: &[String],
-    order_columns: &[String],
-    rowid_order_alias: Option<&str>,
-    limit: usize,
-    offset: usize,
-) -> String {
-    let visible_select_list = if columns.is_empty() {
-        "*".to_string()
-    } else {
-        columns
-            .iter()
-            .map(|column| encode_preview_column_expr(column))
-            .collect::<Vec<_>>()
-            .join(", ")
-    };
-    let order_clause = if order_columns.is_empty() {
-        rowid_order_alias.map_or_else(String::new, |alias| {
-            format!(" ORDER BY {}", quote_ident(alias))
-        })
-    } else {
-        let cols = order_columns
-            .iter()
-            .map(|col| quote_ident(col))
-            .collect::<Vec<_>>()
-            .join(", ");
-        format!(" ORDER BY {cols}")
-    };
-
-    format!(
-        "SELECT {visible_select_list} FROM {}{} LIMIT {} OFFSET {}",
-        quote_ident(table),
-        order_clause,
-        limit,
-        offset
-    )
 }
 
 pub(in crate::adapters::sqlite) fn preview_metadata_query(table: &str) -> String {
@@ -357,51 +318,6 @@ mod tests {
                 "Error: in prepare, no such table: main.pragma_table_list"
             ));
             assert!(!is_table_list_unavailable("FOREIGN KEY constraint failed"));
-        }
-    }
-
-    mod preview_queries {
-        use super::*;
-
-        #[test]
-        fn orders_by_primary_key_columns_when_available() {
-            assert_eq!(
-                build_preview_query(
-                    "users",
-                    &["id".to_string(), "name".to_string()],
-                    &["id".to_string()],
-                    None,
-                    10,
-                    20
-                ),
-                concat!(
-                    r#"SELECT CASE WHEN typeof("id") = 'text' "#,
-                    r#"THEN char(1) || 'SABIQL_HEX:' || hex("id") ELSE "id" END AS "id", "#,
-                    r#"CASE WHEN typeof("name") = 'text' "#,
-                    r#"THEN char(1) || 'SABIQL_HEX:' || hex("name") ELSE "name" END AS "name" "#,
-                    r#"FROM "users" ORDER BY "id" LIMIT 10 OFFSET 20"#
-                )
-            );
-        }
-
-        #[test]
-        fn falls_back_to_star_without_columns() {
-            assert_eq!(
-                build_preview_query("users", &[], &["id".to_string()], None, 10, 20),
-                r#"SELECT * FROM "users" ORDER BY "id" LIMIT 10 OFFSET 20"#
-            );
-        }
-
-        #[test]
-        fn primary_keyless_table_orders_by_rowid_without_selecting_it() {
-            assert_eq!(
-                build_preview_query("logs", &["message".to_string()], &[], Some("rowid"), 10, 0),
-                concat!(
-                    r#"SELECT CASE WHEN typeof("message") = 'text' "#,
-                    r#"THEN char(1) || 'SABIQL_HEX:' || hex("message") ELSE "message" END AS "message" "#,
-                    r#"FROM "logs" ORDER BY "rowid" LIMIT 10 OFFSET 0"#
-                )
-            );
         }
     }
